Extract inner classes to its own files
