+++ lib/fog/dns/google.rb
@@ -1,6 +1,9 @@
 module Fog
   module DNS
     class Google < Fog::Service
+      autoload :Mock, File.expand_path("../google/mock", __FILE__)
+      autoload :Real, File.expand_path("../google/real", __FILE__)
+
       requires :google_project
       recognizes :app_name, :app_version, :google_client_email, :google_key_location, :google_key_string,
                  :google_client, :google_json_key_location, :google_json_key_string
@@ -49,48 +52,6 @@ module Fog
 
       # Project
       request :get_project
-
-      class Mock
-        include Fog::Google::Shared
-
-        def initialize(options)
-          shared_initialize(options[:google_project], GOOGLE_DNS_API_VERSION, GOOGLE_DNS_BASE_URL)
-        end
-
-        def self.data(_api_version)
-          @data ||= {}
-        end
-
-        def self.reset
-          @data = nil
-        end
-
-        def data(project = @project)
-          self.class.data(api_version)[project] ||= {
-            :managed_zones => {},
-            :resource_record_sets => {},
-            :changes => {}
-          }
-        end
-
-        def reset_data
-          self.class.data(api_version).delete(@project)
-        end
-      end
-
-      class Real
-        include Fog::Google::Shared
-
-        attr_accessor :client
-        attr_reader :dns
-
-        def initialize(options)
-          shared_initialize(options[:google_project], GOOGLE_DNS_API_VERSION, GOOGLE_DNS_BASE_URL)
-          options.merge!(:google_api_scope_url => GOOGLE_DNS_API_SCOPE_URLS.join(" "))
-          @client = initialize_google_client(options)
-          @dns = @client.discovered_api("dns", api_version)
-        end
-      end
     end
   end
 end
