+++ lib/fog/google/sql.rb
@@ -1,6 +1,9 @@
 module Fog
   module Google
     class SQL < Fog::Service
+      autoload :Mock, File.expand_path("../sql/mock", __FILE__)
+      autoload :Real, File.expand_path("../sql/real", __FILE__)
+
       requires :google_project
       recognizes :google_client_email, :google_key_location, :google_key_string, :google_client,
                  :app_name, :app_version, :google_json_key_location, :google_json_key_string
@@ -75,56 +78,6 @@ module Fog
 
       # Tier
       request :list_tiers
-
-      class Mock
-        include Fog::Google::Shared
-
-        def initialize(options)
-          shared_initialize(options[:google_project], GOOGLE_SQL_API_VERSION, GOOGLE_SQL_BASE_URL)
-        end
-
-        def self.data
-          @data ||= Hash.new do |hash, key|
-            hash[key] = {
-              :backup_runs => {},
-              :instances => {},
-              :operations => {},
-              :ssl_certs => {}
-            }
-          end
-        end
-
-        def self.reset
-          @data = nil
-        end
-
-        def data
-          self.class.data[project]
-        end
-
-        def reset_data
-          self.class.data.delete(project)
-        end
-
-        def random_operation
-          "operation-#{Fog::Mock.random_numbers(13)}-#{Fog::Mock.random_hex(13)}-#{Fog::Mock.random_hex(8)}"
-        end
-      end
-
-      class Real
-        include Fog::Google::Shared
-
-        attr_accessor :client
-        attr_reader :sql
-
-        def initialize(options)
-          shared_initialize(options[:google_project], GOOGLE_SQL_API_VERSION, GOOGLE_SQL_BASE_URL)
-          options.merge!(:google_api_scope_url => GOOGLE_SQL_API_SCOPE_URLS.join(" "))
-
-          @client = initialize_google_client(options)
-          @sql = @client.discovered_api("sqladmin", api_version)
-        end
-      end
     end
   end
 end
