+++ lib/fog/storage/google_json/mock.rb
@@ -1,0 +1,18 @@
+module Fog
+  module Storage
+    class GoogleJSON
+      class Mock
+        include Utils
+        include Fog::Google::Shared
+
+        def initialize(options = {})
+          shared_initialize(options[:google_project], GOOGLE_STORAGE_JSON_API_VERSION, GOOGLE_STORAGE_JSON_BASE_URL)
+        end
+
+        def signature(_params)
+          "foo"
+        end
+      end
+    end
+  end
+end
