+++ lib/fog/storage/google_json/utils.rb
@@ -1,0 +1,32 @@
+module Fog
+  module Storage
+    class GoogleJSON
+      module Utils
+        def http_url(params, expires)
+          "http://" << host_path_query(params, expires)
+        end
+
+        def https_url(params, expires)
+          "https://" << host_path_query(params, expires)
+        end
+
+        def url(params, expires)
+          Fog::Logger.deprecation("Fog::Storage::Google => #url is deprecated, use #https_url instead [light_black](#{caller.first})[/]")
+          https_url(params, expires)
+        end
+
+        private
+
+        def host_path_query(params, expires)
+          params[:headers]["Date"] = expires.to_i
+          params[:path] = CGI.escape(params[:path]).gsub("%2F", "/")
+          query = [params[:query]].compact
+          query << "GoogleAccessId=#{@client.authorization.issuer}"
+          query << "Signature=#{CGI.escape(signature(params))}"
+          query << "Expires=#{params[:headers]['Date']}"
+          "#{params[:host]}/#{params[:path]}?#{query.join('&')}"
+        end
+      end
+    end
+  end
+end
